pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.42.2",
    "tensorflow>=2.14.0",
]
